% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dates.R
\name{construct_date}
\alias{construct_date}
\title{Construct logical date}
\usage{
construct_date(x, month = 701, day = 15)
}
\arguments{
\item{x}{vector of numerical dates}

\item{month}{replacement for dates with missing month and day}

\item{day}{replacement for dates with missing day}
}
\description{
Transform a numeric value (YYYYMMDD) to a logical numeric date.
The function sets February 29th, on non leap years as Mars 1st.
Dates without months and days are set to mid year of mid month.
}
\examples{
construct_date(18810105)
construct_date(19800000)
construct_date(19810229)

}
