% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dates.R
\name{fill_split}
\alias{fill_split}
\title{Multicore fill missing values by group}
\usage{
fill_split(d, var_names, idvar, up = FALSE)
}
\arguments{
\item{d}{data.frame}

\item{var_names}{vector of variable names}

\item{idvar}{character with name of grouping variable}

\item{up}{boolean direction of fill}
}
\description{
Fast fill in missing values by group. Splits data.frame into groups
and fills missing values using zoo::na.locf.
}
\examples{
df <- data.frame(Month = 1:12, Year = c(2000, rep(NA, 11)), theid = rep(1:2, each = 12))
fill_split(df, "Year", "theid")

}
